+++ ChangeLog.md
@@ -1,6 +1,11 @@
 # Changelog for massaction
 
 ## Unreleased
+- NEW: ajout massaction "Lier commercial" sur la liste des Tiers
+  => donne la possibilité d'ajouter ou de remplacer le ou les commerciaux des Tiers sélectionnés *2021-12-20* - 1.3.0
+- NEW: ajout massaction "Mailing : ajouter destinataires" sur la liste des Tiers, Contacts, Adhérents et Utilisateurs
+=> donne la possibilité d'ajouter des destinataires à un mailing grâce à cette massaction *2021-12-20* - 1.2.0
+- 
 - NEW: cleanup (delete unused files + functions) + compatibility with
   Dolibarr v13-v14 - *2021-08-03* - 1.1.0